Add WgpsMessenger able to conclude commitment scheme
Added a WgpsMessenger which will eventually be able to carry out
a WGPS synchronisation session. Currently it is able to interactively
establish the challenges used to sign when proving ownership of a
secret key.

Added a SyncMessage type.
Added a ControlMessage type.
Added a MsgCommitmentReveal type.

Added MessageEncoder, which encodes messages and queues them for
sending.

Added GrowingBytes, a wrapper for a Uint8Array which is provides methods
for awaiting until the underlying bytes have grown to an absolute size
or a size relative to the current size. This is for the decoders which
need to handle messages as they stream in.

Added decodeMessages generator function which detects which decoder
should be used.

Added decodeCommitmentReveal (streaming)
Added encodeCommitmentReveal (non-streaming)

Add WgpsMessageValidationError for throwing when we get an invalid
message from another peer.

Updated ReadyTransport to take a challenge _hash_ length instead of a
challenge length as one of its params. The passed value can be any
number.